+++ src/main/java/techcourse/myblog/controller/UserController.java
@@ -1,0 +1,10 @@
+package techcourse.myblog.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/users")
+public class UserController {
+
+}
